update build dependencies and config
