+++ package.json
@@ -6,7 +6,8 @@
     "type": "git",
     "url": "git://github.com/tangrams/tangram.git"
   },
-  "main": "dist/tangram.min.js",
+  "main": "dist/tangram.min.mjs",
+  "type": "module",
   "homepage": "https://github.com/tangrams/tangram",
   "keywords": [
     "maps",
@@ -86,6 +87,7 @@
   "license": "MIT",
   "dependencies": {
     "@mapbox/vector-tile": "1.3.0",
+    "core-js": "^3.39.0",
     "csscolorparser": "1.0.3",
     "earcut": "2.2.2",
     "fontfaceobserver": "2.0.7",
@@ -94,36 +96,38 @@
     "gl-mat4": "1.1.4",
     "gl-shader-errors": "1.0.3",
     "js-yaml": "tangrams/js-yaml#read-only",
-    "jszip": "3.1.5",
+    "jszip": "^3.10.1",
     "pbf": "3.1.0",
     "topojson-client": "tangrams/topojson-client#read-only"
   },
   "devDependencies": {
-    "@babel/core": "7.11.4",
-    "@babel/preset-env": "7.11.0",
+    "@babel/core": "^7.26.0",
+    "@babel/plugin-transform-runtime": "^7.25.9",
+    "@babel/preset-env": "^7.26.0",
+    "@rollup/plugin-babel": "^6.0.4",
+    "@rollup/plugin-commonjs": "^28.0.1",
+    "@rollup/plugin-json": "^6.1.0",
+    "@rollup/plugin-node-resolve": "^15.3.0",
+    "@rollup/plugin-replace": "^6.0.1",
+    "@rollup/plugin-terser": "^0.4.4",
+    "@rollup/pluginutils": "^5.1.3",
+    "rollup-plugin-livereload": "^2.0.5",
+    "rollup-plugin-node-builtins": "^2.1.2",
+    "rollup-plugin-node-globals": "^1.4.0",
+    "rollup-plugin-serve": "^1.1.1",
+    "rollup-plugin-sourcemaps2": "^0.4.2",
+    "rollup-plugin-string-import": "^1.2.5",
     "chai": "4.2.0",
     "eslint": "7.8.1",
     "fast-async": "6.3.8",
-    "karma": "4.0.0",
+    "karma": "^6.4.4",
     "karma-chrome-launcher": "2.2.0",
-    "karma-mocha": "1.3.0",
+    "karma-mocha": "^2.0.1",
     "karma-mocha-reporter": "2.2.5",
-    "karma-rollup-preprocessor": "7.0.0-rc.2",
+    "karma-rollup-preprocessor": "^7.0.8",
     "karma-sinon": "1.0.5",
-    "mocha": "5.2.0",
-    "rollup": "1.32.1",
-    "rollup-plugin-babel": "4.4.0",
-    "rollup-plugin-commonjs": "9.2.0",
-    "rollup-plugin-json": "3.1.0",
-    "rollup-plugin-livereload": "1.0.0",
-    "rollup-plugin-node-builtins": "2.1.2",
-    "rollup-plugin-node-globals": "1.4.0",
-    "rollup-plugin-node-resolve": "4.0.0",
-    "rollup-plugin-replace": "2.1.0",
-    "rollup-plugin-serve": "1.0.1",
-    "rollup-plugin-sourcemaps": "0.4.2",
-    "rollup-plugin-string": "2.0.2",
-    "rollup-plugin-terser": "4.0.3",
+    "mocha": "^11.0.1",
+    "rollup": "4.20.0",
     "sinon": "1.10.3"
   }
 }
